Refactor MenuSelectableAnimationProxy
This way everything should work regardless of how it's assigned, assuming there's at most one MSAP on any element
